Rename jinja2 to compose in tests
